create and download csv with incorrect rows
Don't stop the import when finding an error. Let all correct rows to be saved in database and return all error messages after the importer is finished
Insert all wrong/ incorrect rows in a new CSV, saved in local/sepa/import. Every CSV will have a different name and all will be kept.
After the import is finished a "download" button should be displayed, and the user can download the file with incorrect info